Progress on Features Basic login and create account stuff set up (account info is not stored anywhere and the only valid accounts are student and faculty from the Library solution). I made very basic createAccount and login jsp files.
Create poll now uses the servlet rather than the JavaScript. I couldn't
figure out how to preserve the button that was made, but I left the code
for it in so it can be fixed later.

I'll try to work on it more in the next hour and a half, but I'm not
sure how much I can get done.